wyn_load_config: - move k8s client initialization into own file, so watcher subprocesses can have the correct client
